refact(cms): service renaming
