+++ apps/store/.env.development
@@ -1,3 +1,3 @@
-NEXT_PUBLIC_STORYBLOK_URL=https://hedvig.storyblok/api
+NEXT_PUBLIC_CMS_URL=https://hedvig.storyblok/api
 
 NEXT_PUBLIC_API_MOCKING=enabled
