+++ apps/store/.env.local.example
@@ -1,5 +1,5 @@
 # Copy this file as .env.local
-NEXT_PUBLIC_STORYBLOK_URL=
+NEXT_PUBLIC_CMS_URL=
 
 # Optional
 NEXT_PUBLIC_DATADOG_APPLICATION_ID=
