+++ apps/store/src/pages/store.tsx
@@ -3,7 +3,7 @@ import { LayoutWithMenu } from '@/components/LayoutWithMenu/LayoutWithMenu'
 import { StorePage } from '@/components/StorePage/StorePage'
 import { StorePageProps } from '@/components/StorePage/StorePage.types'
 import { getLocale } from '@/lib/l10n/getLocale'
-import { Cms } from '@/services/cms'
+import { CmsService } from '@/services/cms/CmsService'
 
 const NextStorePage: NextPageWithLayout<StorePageProps> = (props: StorePageProps) => {
   return <StorePage {...props} />
@@ -12,7 +12,7 @@ const NextStorePage: NextPageWithLayout<StorePageProps> = (props: StorePageProps
 export const getServerSideProps: GetServerSideProps<StorePageProps> = async (context) => {
   const localeData = getLocale(context.locale ?? context.defaultLocale)
 
-  const products = await Cms.getProductsByMarket(localeData.marketLabel)
+  const products = await CmsService.getProductsByMarket(localeData.marketLabel)
 
   return {
     props: {
